Reorganize chunk pool code structure
