+++ node/chunk_pool/src/mem_pool/mod.rs
@@ -1,0 +1,6 @@
+mod chunk_cache;
+mod chunk_pool_inner;
+mod chunk_write_control;
+
+pub use chunk_pool_inner::MemoryChunkPool;
+pub use chunk_pool_inner::SegmentMetaInfo;
